archive/scripts/requirements.txt
firebase-admin>=6.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
